refactor: split modvendor into download/manifest
